Rename .util.json to .util.parse_json
PyCharm's debugger tries to load our own ".util.json" module instead of
some other internal module, quite possibly because we add a .pch file to
virtualenv's site-packages to make Inline::Python work.
