+++ lib/MediaWords/Controller/Admin/Media.pm
@@ -19,7 +19,7 @@ use MediaWords::DBI::Activities;
 use MediaWords::DBI::Feeds;
 use MediaWords::DBI::Media;
 use MediaWords::DBI::Stories;
-use MediaWords::Util::JSON;
+use MediaWords::Util::ParseJSON;
 use MediaWords::Util::Tags;
 use MediaWords::Util::Web;
 use Readonly;
@@ -141,7 +141,7 @@ sub media_tags_search_json : Local
         )->flat;
     }
 
-    $c->res->body( MediaWords::Util::JSON::encode_json( $terms ) );
+    $c->res->body( MediaWords::Util::ParseJSON::encode_json( $terms ) );
 
     return;
 }
