+++ lib/MediaWords/Controller/Admin/Users.pm
@@ -11,7 +11,7 @@ use parent 'Catalyst::Controller';
 use MediaWords::DBI::Auth;
 use MediaWords::DBI::Auth::Limits;
 use MediaWords::Util::Config;
-use MediaWords::Util::JSON;
+use MediaWords::Util::ParseJSON;
 use MediaWords::Util::Text;
 use POSIX qw(strftime);
 
@@ -490,7 +490,7 @@ sub update_tag_set_permissions_json : Local
         );
     }
 
-    $c->res->body( MediaWords::Util::JSON::encode_json( $tag_set_permissions ) );
+    $c->res->body( MediaWords::Util::ParseJSON::encode_json( $tag_set_permissions ) );
 }
 
 sub tag_set_permissions_json : Local
@@ -516,7 +516,7 @@ sub tag_set_permissions_json : Local
         $userinfo->user_id()
     )->hashes();
 
-    $c->res->body( MediaWords::Util::JSON::encode_json( $auth_users_tag_set_permissions ) );
+    $c->res->body( MediaWords::Util::ParseJSON::encode_json( $auth_users_tag_set_permissions ) );
 }
 
 sub available_tag_sets_json : Local
@@ -541,7 +541,7 @@ sub available_tag_sets_json : Local
         $userinfo->user_id()
     )->hashes();
 
-    $c->res->body( MediaWords::Util::JSON::encode_json( $available_tag_sets ) );
+    $c->res->body( MediaWords::Util::ParseJSON::encode_json( $available_tag_sets ) );
 }
 
 # show the user edit form
@@ -772,7 +772,7 @@ EOF
     }
 
     $c->response->content_type( 'application/json; charset=UTF-8' );
-    $c->response->body( MediaWords::Util::JSON::encode_json( $json_response ) );
+    $c->response->body( MediaWords::Util::ParseJSON::encode_json( $json_response ) );
 }
 
 1;
