+++ lib/MediaWords/Controller/Api/V2/t/Sentences.t
@@ -14,6 +14,8 @@ use MediaWords::Test::DB;
 use MediaWords::Test::Solr;
 use MediaWords::Test::Supervisor;
 
+use MediaWords::Util::ParseJSON;
+
 Readonly my $NUM_MEDIA            => 5;
 Readonly my $NUM_FEEDS_PER_MEDIUM => 2;
 Readonly my $NUM_STORIES_PER_FEED => 10;
@@ -53,7 +55,7 @@ SQL
     my $stories_ids_list = join( ' ', @{ $stories_ids } );
     my $got_ss = test_get( '/api/v2/sentences/list', { q => "stories_id:($stories_ids_list)" } );
 
-    WARN( MediaWords::Util::JSON::encode_json( $got_ss ) );
+    WARN( MediaWords::Util::ParseJSON::encode_json( $got_ss ) );
 
     my $fields = [ qw/stories_id media_id sentence language publish_date/ ];
     rows_match( $label, $got_ss, $expected_ss, 'story_sentences_id', $fields );
