+++ lib/MediaWords/Crawler/Download/Feed/t/univision.t
@@ -19,7 +19,7 @@ use Test::Deep;
 use MediaWords::DB;
 use MediaWords::Test::DB;
 use MediaWords::Crawler::Engine;
-use MediaWords::Util::JSON;
+use MediaWords::Util::ParseJSON;
 use MediaWords::Test::HTTP::HashServer;
 
 sub test_api_request_signature()
@@ -89,7 +89,7 @@ sub test_api_request($$$)
     ok( $json_string, 'JSON response is not empty' );
 
     my $json;
-    eval { $json = MediaWords::Util::JSON::decode_json( $json_string ) };
+    eval { $json = MediaWords::Util::ParseJSON::decode_json( $json_string ) };
     ok( ( !$@ ), "JSON recoding of JSON succeeded: $json_string" );
 
     is( $json->{ 'status' }, 'success', "JSON response was successful: $json_string" );
@@ -205,7 +205,7 @@ sub main()
     say STDERR "Testing against local Univision test HTTP server...";
     my $pages = {
 
-        '/feed' => MediaWords::Util::JSON::encode_json(
+        '/feed' => MediaWords::Util::ParseJSON::encode_json(
             {
                 'status' => 'success',
                 'data'   => {
