+++ lib/MediaWords/Solr.pm
@@ -31,7 +31,7 @@ use MediaWords::Languages::Language;
 use MediaWords::Solr::PseudoQueries;
 use MediaWords::Solr::Query;
 use MediaWords::Util::Config;
-use MediaWords::Util::JSON;
+use MediaWords::Util::ParseJSON;
 use MediaWords::Util::Text;
 use MediaWords::Util::Web;
 
@@ -352,7 +352,7 @@ sub query_encoded_json($$;$)
                 {
                     my $solr_response_json;
 
-                    eval { $solr_response_json = MediaWords::Util::JSON::decode_json( $solr_response_maybe_json ) };
+                    eval { $solr_response_json = MediaWords::Util::ParseJSON::decode_json( $solr_response_maybe_json ) };
                     unless ( $@ )
                     {
                         if (    exists( $solr_response_json->{ error }->{ msg } )
@@ -360,7 +360,8 @@ sub query_encoded_json($$;$)
                         {
                             my $solr_error_msg = $solr_response_json->{ error }->{ msg };
                             my $solr_params =
-                              MediaWords::Util::JSON::encode_json( $solr_response_json->{ responseHeader }->{ params } );
+                              MediaWords::Util::ParseJSON::encode_json(
+                                $solr_response_json->{ responseHeader }->{ params } );
 
                             # If we were able to decode Solr error message, overwrite the default error message with it
                             $error_message = 'Solr error: "' . $solr_error_msg . '", params: ' . $solr_params;
@@ -401,7 +402,7 @@ sub query($$;$)
     my $json = query_encoded_json( $db, $params, $c );
 
     my $data;
-    eval { $data = MediaWords::Util::JSON::decode_json( $json ) };
+    eval { $data = MediaWords::Util::ParseJSON::decode_json( $json ) };
     if ( $@ )
     {
         die( "Error parsing solr json: $@\n$json" );
