+++ lib/MediaWords/t/AbstractJob.t
@@ -6,6 +6,7 @@ use warnings;
 use Modern::Perl "2015";
 use MediaWords::CommonLibs;
 use MediaWords::Test::Supervisor;
+use MediaWords::Util::ParseJSON;
 
 use Sys::Hostname;
 
@@ -97,7 +98,7 @@ sub test_job_state($$$;$)
     is( $js->{ process_id }, $$,                                                    "$label process_id" );
     is( $js->{ priority },   $MediaCloud::JobManager::Job::MJM_JOB_PRIORITY_NORMAL, "$label priority" );
 
-    my $json_data = MediaWords::Util::JSON::decode_json( $js->{ args } );
+    my $json_data = MediaWords::Util::ParseJSON::decode_json( $js->{ args } );
     is( $json_data->{ test }, $test, "$label args JSON test" );
 }
