+++ mediacloud/mediawords/annotator/nyt_labels.py
@@ -2,7 +2,7 @@ from typing import Union, List
 
 from mediawords.annotator import JSONAnnotator, McJSONAnnotatorException
 from mediawords.util.config import get_config as py_get_config
-from mediawords.util.json import encode_json
+from mediawords.util.parse_json import encode_json
 from mediawords.util.log import create_logger
 from mediawords.util.perl import decode_object_from_bytes_if_needed
 from mediawords.util.web.user_agent import Request
