+++ mediacloud/mediawords/util/test_parse_json.py
@@ -1,7 +1,6 @@
 import pytest
 
-from mediawords.util.json import (encode_json, decode_json, McDecodeJSONException,
-                                  McEncodeJSONException)
+from mediawords.util.parse_json import encode_json, decode_json, McDecodeJSONException, McEncodeJSONException
 
 
 def test_encode_decode_json():
