+++ mediacloud/mediawords/util/web/test_user_agent.py
@@ -13,7 +13,7 @@ from furl import furl
 
 from mediawords.test.http.hash_server import HashServer
 from mediawords.util.config import get_config as py_get_config, set_config as py_set_config
-from mediawords.util.json import encode_json, decode_json
+from mediawords.util.parse_json import encode_json, decode_json
 from mediawords.util.log import create_logger
 from mediawords.util.network import random_unused_port
 from mediawords.util.text import random_string
